# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notice/search": {
      "get": {
        "operationId": "NoticeController_search",
        "parameters": [
          {
            "name": "pageNo",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NoticeResponseDTO"
                      }
                    },
                    "pageInfo": {
                      "$ref": "#/components/schemas/PageInfoDTO"
                    }
                  },
                  "required": [
                    "items",
                    "pageInfo"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "notice"
        ]
      }
    },
    "/api/notice/{id}": {
      "get": {
        "operationId": "NoticeController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "tags": [
          "notice"
        ]
      },
      "patch": {
        "operationId": "NoticeController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoticeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notice"
        ]
      },
      "delete": {
        "operationId": "NoticeController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notice"
        ]
      }
    },
    "/api/notice": {
      "post": {
        "operationId": "NoticeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoticeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notice"
        ]
      }
    },
    "/api/notice/{id}/pin": {
      "patch": {
        "operationId": "NoticeController_togglePin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notice"
        ]
      }
    },
    "/api/notice/{id}/expose": {
      "patch": {
        "operationId": "NoticeController_toggleExpose",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticeResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notice"
        ]
      }
    },
    "/api/faq/search": {
      "get": {
        "operationId": "FaqController_search",
        "parameters": [
          {
            "name": "pageNo",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FaqResponseDTO"
                      }
                    },
                    "pageInfo": {
                      "$ref": "#/components/schemas/PageInfoDTO"
                    }
                  },
                  "required": [
                    "items",
                    "pageInfo"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "faq"
        ]
      }
    },
    "/api/faq/{id}": {
      "get": {
        "operationId": "FaqController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "tags": [
          "faq"
        ]
      },
      "patch": {
        "operationId": "FaqController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFaqDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "faq"
        ]
      },
      "delete": {
        "operationId": "FaqController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "faq"
        ]
      }
    },
    "/api/faq": {
      "post": {
        "operationId": "FaqController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFaqDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "faq"
        ]
      }
    },
    "/api/faq/{id}/pin": {
      "patch": {
        "operationId": "FaqController_togglePin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "faq"
        ]
      }
    },
    "/api/faq/{id}/expose": {
      "patch": {
        "operationId": "FaqController_toggleExpose",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaqResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "faq"
        ]
      }
    },
...
  },
  "info": {
    "title": "Demo Shop API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageInfoDTO": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "pageNo",
          "pageSize",
          "totalItems",
          "totalPages"
        ]
      },
      "NoticeResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "isExposed": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "imageURL": {
            "type": "string",
            "nullable": true
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "isExposed",
          "isPinned",
          "imageURL",
          "fileUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateNoticeDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "imageURL": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "UpdateNoticeDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "imageURL": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          }
        }
      },
      "FaqResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "isExposed": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "question",
          "answer",
          "isExposed",
          "isPinned",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateFaqDTO": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "question",
          "answer"
        ]
      },
      "UpdateFaqDTO": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          }
        }
      },
...
    }
  }
}